review: refactor to move calendarHandler outside subscribe handler
