+++ packages/server-common/src/handlers/calendar/index.ts
@@ -1,3 +1,4 @@
+export * from './calendar-handler-factory.contentful';
+export * from './calendar-handler-factory.squidex';
 export * from './resubscribe-handler';
 export * from './subscribe-handler';
-export * from './subscribe-handler.contentful';
